refactor Mineral, elastic and stress
