--- .readthedocs.yml
@@ -1,14 +1,0 @@
-# Required
-version: 2
-
-# Image to use
-build:
-  image: testing
-
-# Configuration
-python:
-  version: 3.9
-  install:
-    - requirements: docs/requirements.txt
-    - method: pip
-      path: .
