+++ docs/api/base_classes.md
@@ -6,3 +6,18 @@
       show_source: false
       heading_level: 3
 
+## `digirock.Fluid`
+::: digirock.Fluid
+    rendering:
+      show_root_heading: false
+      show_root_toc_entry: false
+      show_source: false
+      heading_level: 3
+
+## `digirock.StressModel`
+::: digirock.StressModel
+    rendering:
+      show_root_heading: false
+      show_root_toc_entry: false
+      show_source: false
+      heading_level: 3
