+++ docs/api/elastic.md
@@ -1,0 +1,9 @@
+## Elastic Methods
+
+```python
+from digirock.elastic import *
+```
+
+::: digirock.elastic
+    rendering:
+      heading_level: 3
