+++ docs/api/fluid_classes.md
@@ -1,11 +1,3 @@
-## `digirock.Fluid`
-::: digirock.Fluid
-    rendering:
-      show_root_heading: false
-      show_root_toc_entry: false
-      show_source: false
-      heading_level: 3
-
 ## `digirock.Water`
 ::: digirock.Water
     rendering:
