+++ docs/api/minerals.md
@@ -1,0 +1,7 @@
+## `digirock.Mineral`
+::: digirock.Mineral
+    rendering:
+      show_root_heading: false
+      show_root_toc_entry: false
+      show_source: false
+      heading_level: 3
