+++ docs/api/stress_classes.md
@@ -1,0 +1,15 @@
+## `digirock.FStressModel`
+::: digirock.FStressModel
+    rendering:
+      show_root_heading: false
+      show_root_toc_entry: false
+      show_source: false
+      heading_level: 3
+
+## `digirock.LGStressModel`
+::: digirock.LGStressModel
+    rendering:
+      show_root_heading: false
+      show_root_toc_entry: false
+      show_source: false
+      heading_level: 3
