+++ setup.cfg
@@ -23,6 +23,8 @@ install_requires =
     scipy
     numba
     addict
+    more_itertools
+    rich
 
 [options.packages.find]
 where = src
@@ -34,6 +36,7 @@ test =
     pandas
     numpy
     xarray
+    hypothesis
 
 docs =
     mkdocs
@@ -44,3 +47,4 @@ docs =
     mkdocs-jupyter
     mkautodoc
     jupyterlab
+    matplotlib
